(refactor) Get rid of SettingsForm.vue
